fix: 리뷰 컨트롤러 테스트 버그 해결 createdDate의 기본값을 현재 시간으로 하여 Nullable=false를 지키면서 App에서 객체 생성이 가능하다.

+++ src/main/java/bdbe/bdbd/review/ReviewResponse.java
@@ -10,7 +10,7 @@ public class ReviewResponse {
     @Setter
     public static class getReviewById{
         private Long id;
-        private int uId;
+        private Long uId;
         private Long cId;
         private Reservation reservation; //1:1, 참조용(read-only)
         private String comment;
